Fix/update tests
