+++ test/project/tests/__init__.py
@@ -1,3 +1,4 @@
 from .types import *
-from .arg_types import *
 from .gdextension_api import *
+from .arg_types import *
+from .arg_types_ext import *
